Some initial graph based models ideas.
The end game is not to take in a list of Dbee::Base objects. However, I wanted to start from this point of view as it most closely aligns with the graph based approach to modeling. There are obvious issus with this approach with the most obvious being that the Dbee::Base class names to not match the names of the associations.